Create notification service
Add a new microservice that sends mails to the admin.
Created a scheduler but also an endpoint to send metrics.

TO DO:
Send daily mail with top medias (recommended)
Send daily with metrics
Send alerts when reaching a certain value
Send mail when uploading a file
Format mail add a picture
see the meaning of each metrics